Renamed in line with the `imagerotator` service
