+++ stack/services/imagegrayscale/.gitignore
