+++ stack/services/imagegrayscale/build.gradle
