+++ stack/services/imagegrayscale/gradle/wrapper/gradle-wrapper.jar
