+++ stack/services/imagegrayscale/gradle/wrapper/gradle-wrapper.properties
