+++ stack/services/imagegrayscale/gradlew
