+++ stack/services/imagegrayscale/gradlew.bat
