+++ stack/services/imagegrayscale/settings.gradle
