+++ stack/services/imagegrayscale/src/main/java/com/github/olly/workshop/imagegrayscale/ImageGrayscaleApplication.java
@@ -1,13 +1,13 @@
-package com.github.olly.workshop.imagetransformer;
+package com.github.olly.workshop.imagegrayscale;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ImagetransformerApplication {
+public class ImageGrayscaleApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ImagetransformerApplication.class, args);
+		SpringApplication.run(ImageGrayscaleApplication.class, args);
 	}
 
 }
