+++ stack/services/imagegrayscale/src/main/java/com/github/olly/workshop/imagegrayscale/service/ImageService.java
@@ -1,4 +1,4 @@
-package com.github.olly.workshop.imagetransformer.service;
+package com.github.olly.workshop.imagegrayscale.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
