+++ stack/services/imagegrayscale/src/main/resources/application.properties
@@ -5,5 +5,5 @@ management.metrics.export.prometheus.enabled=true
 spring.servlet.multipart.max-file-size:10MB
 spring.servlet.multipart.max-request-size:10MB
 logging.pattern.console=%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} [%X{traceId}/%X{spanId}] - %msg %replace(%xException){'\n','\r'}%nopex%n
-spring.application.name=imagetransformer
+spring.application.name=imagegrayscale
 server.port=8181
